improve(pinia): move fio to planning store
also removes now unused gameDataStore (IndexedDB now) references + tests
